ruruby-parse: Refactor.
