+++ bin/test
@@ -1,5 +1,4 @@
 #!/bin/bash
-set -x
 cargo clean
 cargo install --features gc-stress --path monoruby
