+++ ruruby-parse/src/parser/expression.rs
@@ -450,7 +450,7 @@ impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
             let loc = self.prev_loc();
             match self.peek_no_term()?.kind {
                 TokenKind::IntegerLit(_) | TokenKind::FloatLit(_) | TokenKind::BignumLit(_) => {
-                    self.parse_primary(true)?;
+                    let _ = self.parse_primary()?;
                     if self.consume_punct_no_term(Punct::DMul)? {
                         self.restore_state(save2);
                         let lhs = self.parse_exponent()?;
@@ -543,7 +543,7 @@ impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
         // 一次式メソッド呼び出し
         // スコープ付き定数参照 :: 一次式 [行終端子禁止][空白類禁止] "::" 定数識別子
         //      ｜"::" 定数識別子
-        let mut node = self.parse_primary(false)?;
+        let mut node = self.parse_primary()?;
         loop {
             node = if self.consume_punct(Punct::Dot)? {
                 self.parse_primary_method(node, false)?
@@ -553,9 +553,8 @@ impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
                 if let TokenKind::Const(_) = self.peek()?.kind {
                     let loc = node.loc;
                     let name = self.expect_const()?;
-                    if self.consume_punct_no_term(Punct::LParen)? {
+                    if let Some(arglist) = self.parse_arguments_paren(true)? {
                         // Foo::Bar()
-                        let arglist = self.parse_arglist_block(Punct::RParen)?;
                         Node::new_mcall(node, name, arglist, false, self.prev_loc())
                     } else if let NodeKind::Const {
                         toplevel,
@@ -600,28 +599,19 @@ impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
         {
             return Ok(Node::new_yield(ArgList::default(), loc));
         };
-        let args = if self.consume_punct(Punct::LParen)? {
-            self.parse_arglist(Punct::RParen)?
+        let arglist = if let Some(arglist) = self.parse_arguments(false)? {
+            arglist
         } else {
-            self.parse_arglist(None)?
+            ArgList::default()
         };
-        Ok(Node::new_yield(args, loc))
-    }
-
-    fn parse_super(&mut self) -> Result<Node, LexerErr> {
-        let loc = self.prev_loc();
-        let arglist = if self.consume_punct_no_term(Punct::LParen)? {
-            self.parse_arglist_block(Punct::RParen)?
-        } else if self.is_command() {
-            self.parse_arglist_block(None)?
-        } else {
-            return Ok(Node::new_super(None, loc));
-        };
-        let loc = self.prev_loc().merge(loc);
-        Ok(Node::new_super(arglist, loc))
+        Ok(Node::new_yield(arglist, loc))
     }
 
     /// Parse primary method call.
+    ///
+    /// `receiver`.
+    /// `receiver`?
+    /// `receiver`::
     pub(super) fn parse_primary_method(
         &mut self,
         receiver: Node,
@@ -636,377 +626,30 @@ impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
         //      ｜ 一次式 ［行終端子禁止］ "::" メソッド名 括弧付き実引数 ブロック?
         //      ｜ 一次式 ［行終端子禁止］ "::" 定数以外のメソッド名 ブロック?
         let loc = self.prev_loc;
-        if self.consume_punct_no_term(Punct::LParen)? {
-            let arglist = self.parse_arglist_block(Punct::RParen)?;
-            //let loc = receiver.loc().merge(self.loc());
+        if let Some(arglist) = self.parse_arguments_paren(true)? {
+            // Support https://docs.ruby-lang.org/ja/latest/doc/spec=2fcall.html#call_method
+            // `receiver`.(..)
+            // `receiver`?(..)
+            // `receiver`::(..)
             let node = Node::new_mcall(receiver, "call".to_string(), arglist, false, loc);
             return Ok(node);
         };
         let (id, loc) = self.read_method_name(false)?;
-        let arglist = if self.consume_punct_no_term(Punct::LParen)? {
-            self.parse_arglist_block(Punct::RParen)?
+        let arglist = if let Some(arglist) = self.parse_arguments(true)? {
+            // `receiver`.method(..)
+            arglist
         } else {
-            if self.is_command() {
-                return Ok(Node::new_mcall(
-                    receiver,
-                    id,
-                    self.parse_arglist_block(None)?,
-                    false,
-                    loc,
-                ));
-            }
             match self.parse_block()? {
                 Some(block) => ArgList::with_block(block),
                 None => ArgList::default(),
             }
         };
 
-        let node = match receiver.kind {
+        let receiver = match receiver.kind {
             NodeKind::Ident(id) => Node::new_fcall_noarg(id, false, loc),
             _ => receiver,
         };
-        Ok(Node::new_mcall(node, id, arglist, safe_nav, loc))
-    }
-
-    pub(super) fn parse_primary(&mut self, suppress_unparen_call: bool) -> Result<Node, LexerErr> {
-        let tok = self.get()?;
-        let loc = tok.loc();
-        match tok.kind {
-            TokenKind::Ident(name) => {
-                match name.as_str() {
-                    "true" => return Ok(Node::new_bool(true, loc)),
-                    "false" => return Ok(Node::new_bool(false, loc)),
-                    "nil" => return Ok(Node::new_nil(loc)),
-                    "self" => return Ok(Node::new_self(loc)),
-                    "__LINE__" => {
-                        let line = self.lexer.get_line(loc.0);
-                        return Ok(Node::new_integer(line as i64, loc));
-                    }
-                    "__FILE__" => {
-                        let file = self.path.to_string_lossy().to_string();
-                        return Ok(Node::new_string(file.into(), loc));
-                    }
-                    _ => {}
-                };
-
-                if self.lexer.trailing_lparen() {
-                    let node = Node::new_identifier(name.clone(), loc);
-                    return self.parse_function_args(node);
-                };
-                if let Some(outer) = self.is_local_var(&name) {
-                    Ok(Node::new_lvar(name, outer, loc))
-                } else {
-                    // FUNCTION or COMMAND or LHS for assignment
-                    let node = Node::new_identifier(name.to_string(), loc);
-                    if let Ok(tok) = self.peek_no_term() {
-                        match tok.kind {
-                            // Multiple assignment
-                            TokenKind::Punct(Punct::Comma) => return Ok(node),
-                            // Method call with block and no args
-                            TokenKind::Punct(Punct::LBrace) | TokenKind::Reserved(Reserved::Do) => {
-                                return self.parse_function_args(node)
-                            }
-                            _ => {}
-                        }
-                    };
-
-                    if !suppress_unparen_call && self.is_command() {
-                        Ok(self.parse_command(name, loc)?)
-                    } else {
-                        Ok(node)
-                    }
-                }
-            }
-            TokenKind::NumberedParam(i, name) => {
-                if self.lexer.trailing_lparen() {
-                    // _1()
-                    let node = Node::new_identifier(name.clone(), loc);
-                    return self.parse_function_args(node);
-                };
-                self.check_outer_numbered_param(loc)?;
-                // FUNCTION or COMMAND or LHS for assignment
-                let node = Node::new_identifier(name.to_string(), loc);
-                if let Ok(tok) = self.peek_no_term() {
-                    match tok.kind {
-                        // Multiple assignment
-                        TokenKind::Punct(Punct::Comma) => return Err(error_numbered_param(loc, i)),
-                        // Method call with block and no args
-                        // _1 {}
-                        TokenKind::Punct(Punct::LBrace) | TokenKind::Reserved(Reserved::Do) => {
-                            return self.parse_function_args(node)
-                        }
-                        _ => {}
-                    }
-                };
-
-                if !suppress_unparen_call && self.is_command() {
-                    Ok(self.parse_command(name, loc)?)
-                } else {
-                    Ok(node)
-                }
-            }
-            TokenKind::InstanceVar(name) => Ok(Node::new_instance_var(name, loc)),
-            TokenKind::ClassVar(name) => Ok(Node::new_class_var(name, loc)),
-            TokenKind::GlobalVar(name) => Ok(Node::new_global_var(name, loc)),
-            TokenKind::SpecialVar(id) => Ok(Node::new_special_var(id, loc)),
-            TokenKind::Const(name) => {
-                if self.lexer.trailing_lparen() {
-                    let node = Node::new_identifier(name, loc);
-                    self.parse_function_args(node)
-                } else if !suppress_unparen_call && self.is_command() {
-                    Ok(self.parse_command(name, loc)?)
-                } else {
-                    Ok(Node::new_const(name, false, None, vec![], loc))
-                }
-            }
-            TokenKind::IntegerLit(num) => Ok(Node::new_integer(num, loc)),
-            TokenKind::BignumLit(num) => Ok(Node::new_bignum(num, loc)),
-            TokenKind::FloatLit(num) => Ok(Node::new_float(num, loc)),
-            TokenKind::ImaginaryLit(num) => Ok(Node::new_imaginary(num, loc)),
-            TokenKind::StringLit(s) => self.parse_string_literal(s),
-            TokenKind::CommandLit(s) => {
-                let content = Node::new_string(s.into(), loc);
-                Ok(Node::new_command(content))
-            }
-            TokenKind::OpenString(s, term, level) => {
-                self.parse_interporated_string_literal(s.into(), term, level)
-            }
-            TokenKind::OpenCommand(s, term, level) => {
-                let content = self.parse_interporated_string_literal(s.into(), term, level)?;
-                Ok(Node::new_command(content))
-            }
-            TokenKind::Punct(punct) => match punct {
-                Punct::Minus => match self.get()?.kind {
-                    TokenKind::IntegerLit(num) => match num.checked_neg() {
-                        Some(i) => Ok(Node::new_integer(i, loc)),
-                        None => Ok(Node::new_bignum(-BigInt::from(num), loc)),
-                    },
-                    TokenKind::BignumLit(num) => Ok(Node::new_bignum(-num, loc)),
-                    TokenKind::FloatLit(num) => Ok(Node::new_float(-num, loc)),
-                    _ => unreachable!(),
-                },
-                Punct::LParen => {
-                    let old = self.suppress_mul_assign;
-                    self.suppress_mul_assign = false;
-                    let node = self.parse_comp_stmt()?;
-                    self.expect_punct(Punct::RParen)?;
-                    self.suppress_mul_assign = old;
-                    Ok(node)
-                }
-                Punct::LBracket => {
-                    // Array literal
-                    let nodes = self.parse_mul_assign_rhs(Punct::RBracket, true)?;
-                    let loc = loc.merge(self.prev_loc());
-                    Ok(Node::new_array(nodes, loc))
-                }
-                Punct::LBrace => self.parse_hash_literal(false, None),
-                Punct::Colon => self.parse_symbol(),
-                Punct::Arrow => self.parse_lambda_literal(),
-                Punct::Scope => {
-                    let name = self.expect_const()?;
-                    Ok(Node::new_const(name, true, None, vec![], loc))
-                }
-                Punct::Div => self.parse_regexp(),
-                Punct::Rem => self.parse_percent_notation(),
-                Punct::Question => self.parse_char_literal(),
-                Punct::Shl => self.parse_heredocument(),
-                Punct::Mul => {
-                    let first_lhs = self.parse_method_call()?;
-                    let loc = loc.merge(first_lhs.loc());
-                    let first_lhs = Node::new_splat(first_lhs, loc);
-                    if self.consume_punct_no_term(Punct::Comma)? {
-                        self.parse_mul_assign(first_lhs)
-                    } else if self.consume_punct_no_term(Punct::Assign)? {
-                        let mrhs = self.parse_mul_assign_rhs_if_allowed()?;
-                        let lhs = self.check_lhs(first_lhs)?;
-                        Ok(Node::new_mul_assign(vec![lhs], mrhs))
-                    } else {
-                        let loc = self.loc();
-                        return Err(error_unexpected(loc, "Expected '=' or ','."));
-                    }
-                }
-                _ => Err(error_unexpected(
-                    loc,
-                    format!("Unexpected token: {:?}", tok.kind),
-                )),
-            },
-            TokenKind::Reserved(reserved) => match reserved {
-                Reserved::If => self.parse_if(),
-                Reserved::Unless => self.parse_unless(),
-                Reserved::For => self.parse_for(),
-                Reserved::While => self.parse_while(true),
-                Reserved::Until => self.parse_while(false),
-                Reserved::Case => self.parse_case(),
-                Reserved::Def => self.parse_def(),
-                Reserved::Class => {
-                    let loc = self.prev_loc();
-                    if self.consume_punct(Punct::Shl)? {
-                        self.parse_singleton_class(loc)
-                    } else {
-                        if self.is_method_context() {
-                            return Err(error_unexpected(
-                                loc,
-                                "SyntaxError: class definition in method body.",
-                            ));
-                        }
-                        self.parse_class(false)
-                    }
-                }
-                Reserved::Module => {
-                    if self.is_method_context() {
-                        return Err(error_unexpected(
-                            loc,
-                            "SyntaxError: module definition in method body.",
-                        ));
-                    }
-                    self.parse_class(true)
-                }
-                Reserved::Return => self.parse_return(),
-                Reserved::Break => self.parse_break(),
-                Reserved::Next => self.parse_next(),
-                Reserved::Redo => self.parse_redo(),
-                Reserved::Begin => self.parse_begin(),
-                Reserved::Defined => {
-                    if self.consume_punct_no_term(Punct::LParen)? {
-                        self.defined_mode = true;
-                        let node = self.parse_expr()?;
-                        self.defined_mode = false;
-                        self.expect_punct(Punct::RParen)?;
-                        Ok(Node::new_defined(node))
-                    } else {
-                        let tok = self.get()?;
-                        Err(error_unexpected(tok.loc, "expected '('.".to_string()))
-                    }
-                }
-                Reserved::Alias => {
-                    let new_name = self.alias_name()?;
-                    let old_name = self.alias_name()?;
-                    let loc = loc.merge(self.prev_loc());
-                    Ok(Node::new_alias(new_name, old_name, loc))
-                }
-                Reserved::Super => self.parse_super(),
-                _ => Err(error_unexpected(
-                    loc,
-                    format!("Unexpected token: {:?}", tok.kind),
-                )),
-            },
-            TokenKind::Eof => Err(error_eof(loc)),
-            _ => Err(error_unexpected(
-                loc,
-                format!("Unexpected token: {:?}", tok.kind),
-            )),
-        }
-    }
-
-    fn parse_arglist_block(
-        &mut self,
-        delimiter: impl Into<Option<Punct>>,
-    ) -> Result<ArgList, LexerErr> {
-        let mut arglist = self.parse_arglist(delimiter)?;
-        if let Some(actual_block) = self.parse_block()? {
-            if arglist.block.is_some() {
-                return Err(error_unexpected(
-                    actual_block.loc(),
-                    "Both block arg and actual block given.",
-                ));
-            }
-            arglist.block = Some(actual_block);
-        };
-        Ok(arglist)
-    }
-
-    /// Parse argument list.
-    /// arg, *splat_arg, kw: kw_arg, **double_splat_arg, &block <punct>
-    /// punct: punctuator for terminating arg list. Set None for unparenthesized argument list.
-    fn parse_arglist(&mut self, punct: impl Into<Option<Punct>>) -> Result<ArgList, LexerErr> {
-        let punct = punct.into();
-        let mut arglist = ArgList::default();
-        if self.peek()?.kind == TokenKind::Punct(Punct::Comma) {
-            return Ok(arglist);
-        }
-        loop {
-            if let Some(punct) = punct {
-                if self.consume_punct(punct)? {
-                    return Ok(arglist);
-                }
-            }
-            if self.consume_punct(Punct::Range3)? {
-                self.check_forwarding()?;
-                arglist.forwarding = true;
-            } else if self.consume_punct(Punct::Mul)? {
-                // splat argument
-                let loc = self.prev_loc();
-                let array = self.parse_arg(false)?;
-                arglist.splat = true;
-                arglist.args.push(Node::new_splat(array, loc));
-            } else if self.consume_punct(Punct::DMul)? {
-                // double splat argument
-                arglist.hash_splat.push(self.parse_arg(false)?);
-            } else if self.consume_punct(Punct::BitAnd)? {
-                // block argument
-                arglist.block = Some(Box::new(self.parse_arg(false)?));
-            } else {
-                let node = self.parse_arg(false)?;
-                let loc = node.loc();
-                if self.consume_punct(Punct::FatArrow)? {
-                    let value = self.parse_arg(false)?;
-                    let mut kvp = vec![(node, value)];
-                    if self.consume_punct(Punct::Comma)? {
-                        loop {
-                            // Support trailing comma
-                            if let Some(punct) = punct {
-                                if let TokenKind::Punct(p) = self.peek()?.kind {
-                                    if punct == p {
-                                        break;
-                                    }
-                                }
-                            };
-                            let key = self.parse_arg(false)?;
-                            self.expect_punct(Punct::FatArrow)?;
-                            let value = self.parse_arg(false)?;
-                            kvp.push((key, value));
-                            if !self.consume_punct(Punct::Comma)? {
-                                break;
-                            }
-                        }
-                    }
-                    if let Some(punct) = punct {
-                        self.consume_punct(punct)?;
-                    };
-                    let node = Node::new_hash(kvp, loc);
-                    arglist.args.push(node);
-                    return Ok(arglist);
-                }
-                match &node.kind {
-                    NodeKind::Ident(id, ..) | NodeKind::LocalVar(0, id) => {
-                        if self.consume_punct_no_term(Punct::Colon)? {
-                            // keyword args
-                            arglist
-                                .kw_args
-                                .push((id.to_string(), self.parse_arg(false)?));
-                        } else {
-                            // positional args
-                            arglist.args.push(node);
-                        }
-                    }
-                    _ => {
-                        arglist.args.push(node);
-                    }
-                }
-            }
-            if !self.consume_punct(Punct::Comma)? {
-                break;
-            } else {
-                let loc = self.prev_loc();
-                if arglist.block.is_some() {
-                    return Err(error_unexpected(loc, "unexpected ','."));
-                };
-            }
-        }
-        if let Some(punct) = punct {
-            self.consume_punct(punct)?;
-        };
-        Ok(arglist)
+        Ok(Node::new_mcall(receiver, id, arglist, safe_nav, loc))
     }
 
     pub(super) fn parse_begin(&mut self) -> Result<Node, LexerErr> {
@@ -1050,35 +693,19 @@ impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
         Ok(Node::new_begin(body, rescue, else_, ensure))
     }
 
-    fn parse_command(&mut self, operation: String, loc: Loc) -> Result<Node, LexerErr> {
-        // FNAME ARGS
-        // FNAME ARGS DO-BLOCK
-        let send_args = self.parse_arglist_block(None)?;
-        Ok(Node::new_fcall(operation, send_args, false, loc))
-    }
-
-    fn parse_function_args(&mut self, node: Node) -> Result<Node, LexerErr> {
-        let loc = node.loc();
-        if self.consume_punct_no_term(Punct::LParen)? {
-            // PRIMARY-METHOD : FNAME ( ARGS ) BLOCK?
-            let send_args = self.parse_arglist_block(Punct::RParen)?;
-
-            Ok(Node::new_fcall(
-                node.as_method_name().unwrap(),
-                send_args,
-                false,
-                loc,
-            ))
-        } else if let Some(block) = self.parse_block()? {
-            // PRIMARY-METHOD : FNAME BLOCK
-            Ok(Node::new_fcall(
-                node.as_method_name().unwrap(),
-                ArgList::with_block(block),
-                false,
-                loc,
-            ))
+    pub(super) fn parse_splat(&mut self, loc: Loc) -> Result<Node, LexerErr> {
+        let first_lhs = self.parse_method_call()?;
+        let loc = loc.merge(first_lhs.loc());
+        let first_lhs = Node::new_splat(first_lhs, loc);
+        if self.consume_punct_no_term(Punct::Comma)? {
+            self.parse_mul_assign(first_lhs)
+        } else if self.consume_punct_no_term(Punct::Assign)? {
+            let mrhs = self.parse_mul_assign_rhs_if_allowed()?;
+            let lhs = self.check_lhs(first_lhs)?;
+            Ok(Node::new_mul_assign(vec![lhs], mrhs))
         } else {
-            Ok(node)
+            let loc = self.loc();
+            return Err(error_unexpected(loc, "Expected '=' or ','."));
         }
     }
 
@@ -1115,55 +742,5 @@ impl<'a, OuterContext: LocalsContext> Parser<'a, OuterContext> {
             return Ok(Node::new_splat(self.check_lhs(*lhs)?, loc));
         }
         Ok(lhs)
-    }
-
-    fn is_command(&mut self) -> bool {
-        let tok = match self.peek_no_term() {
-            Ok(tok) => tok,
-            _ => return false,
-        };
-        if self.lexer.trailing_space() {
-            match tok.kind {
-                TokenKind::LineTerm => false,
-                TokenKind::Punct(p) => match p {
-                    Punct::LParen | Punct::LBracket | Punct::Scope | Punct::Arrow | Punct::Not => {
-                        true
-                    }
-                    Punct::Colon
-                    | Punct::Plus
-                    | Punct::Minus
-                    | Punct::Mul
-                    | Punct::Div
-                    | Punct::BitAnd
-                    | Punct::Rem
-                    | Punct::Shl => !self.lexer.has_trailing_space(&tok),
-                    _ => false,
-                },
-                TokenKind::Reserved(r) => !matches!(
-                    r,
-                    Reserved::Do
-                        | Reserved::If
-                        | Reserved::Unless
-                        | Reserved::Rescue
-                        | Reserved::While
-                        | Reserved::Until
-                        | Reserved::And
-                        | Reserved::Or
-                        | Reserved::Then
-                        | Reserved::End
-                ),
-                _ => true,
-            }
-        } else {
-            matches!(
-                tok.kind,
-                TokenKind::GlobalVar(_)
-                    | TokenKind::InstanceVar(_)
-                    | TokenKind::StringLit(_)
-                    | TokenKind::FloatLit(_)
-                    | TokenKind::BignumLit(_)
-                    | TokenKind::IntegerLit(_)
-            )
-        }
     }
 }
